OpenGDC update 20180309

+++ src/opengdc/action/DownloadDataAction.java
@@ -42,7 +42,7 @@ public class DownloadDataAction extends Action {
         
         // create log window
         logPane = GUI.createLogWindow();
-        
+        System.out.println("DOWNLOAD\tDisease: "+disease+"\tDataType: "+dataType);
         System.err.println("Downloading GDC Data" + "\n" + "Disease: " + disease + "\n" + "Data Type: " + dataType + "\n" + "Output Folder Path: " + gdc_path + "\n" + "Auto-extract: " + autoextract + "\n" + "Auto-remove: " + autoremove + "\n");
         GUI.appendLog(logPane, "Downloading GDC Data" + "\n" + "Disease: " + disease + "\n" + "Data Type: " + dataType + "\n" + "Output Folder Path: " + gdc_path + "\n" + "Auto-extract: " + autoextract + "\n" + "Auto-remove: " + autoremove + "\n");
         
@@ -70,48 +70,52 @@ public class DownloadDataAction extends Action {
         for (String uuid: dataMap.keySet()) {
             // download data
             String fileName = uuid + "_" + dataMap.get(uuid).get("file_name");
-            for (String s: dataMap.get(uuid).keySet())
-                System.err.println(s + "\t" + dataMap.get(uuid).get(s));
-            
-            GDCQuery.downloadFile(uuid, gdc_path, fileName, false, 0, logPane);
-            if (autoextract) {
-                HashSet<String> uncompressed_folders_path = new HashSet<>();
-                HashSet<String> experiments_path = new HashSet<>();
-                File data_file = new File(gdc_path + fileName);
-                if (data_file.exists()) {
-                    String destDirPath = gdc_path + "/" + data_file.getName() + "_/";
-                    File destDir = new File(destDirPath);
-                    
-                    // extract data
-                    boolean uncompressed = DataExtractionTool.uncompressData(data_file, destDir, false, true);
-                    //System.out.println("uncompressed: " + uncompressed);
-                    uncompressed_folders_path.addAll(DataExtractionTool.getUncompressedFoldersPathList());
-                    experiments_path.addAll(DataExtractionTool.getExperimentsPathList());
-                    
-                    // copy experiments to the gdc_path folder
-                    for (String exp: experiments_path) {
-                        try {
-                            File expFile = new File(exp);
-                            File newExpFile = new File(gdc_path + expFile.getName());
-                            Files.move(expFile.toPath(), newExpFile.toPath(), REPLACE_EXISTING);
+            File data_file = new File(gdc_path + "/" + fileName);
+            if (!data_file.exists()) {
+                for (String s: dataMap.get(uuid).keySet())
+                    System.err.println(s + "\t" + dataMap.get(uuid).get(s));
+
+                GDCQuery.downloadFile(uuid, gdc_path, fileName, false, 0, logPane);
+                if (autoextract) {
+                    HashSet<String> uncompressed_folders_path = new HashSet<>();
+                    HashSet<String> experiments_path = new HashSet<>();
+                    if (data_file.exists()) {
+                        String destDirPath = gdc_path + "/" + data_file.getName() + "_/";
+                        File destDir = new File(destDirPath);
+
+                        // extract data
+                        boolean uncompressed = DataExtractionTool.uncompressData(data_file, destDir, false, true);
+                        //System.out.println("uncompressed: " + uncompressed);
+                        uncompressed_folders_path.addAll(DataExtractionTool.getUncompressedFoldersPathList());
+                        experiments_path.addAll(DataExtractionTool.getExperimentsPathList());
+
+                        // copy experiments to the gdc_path folder
+                        for (String exp: experiments_path) {
+                            try {
+                                File expFile = new File(exp);
+                                File newExpFile = new File(gdc_path + expFile.getName());
+                                Files.move(expFile.toPath(), newExpFile.toPath(), REPLACE_EXISTING);
+                            }
+                            catch (Exception e) {
+                                //e.printStackTrace();
+                            }
                         }
-                        catch (Exception e) {
-                            //e.printStackTrace();
+
+                        // remove other files and folders
+                        for (String dir: uncompressed_folders_path) {
+                            File dirFolder = new File(dir);
+                            FSUtils.deleteDir(dirFolder);
                         }
-                    }
-                    
-                    // remove other files and folders
-                    for (String dir: uncompressed_folders_path) {
-                        File dirFolder = new File(dir);
-                        FSUtils.deleteDir(dirFolder);
-                    }
-                    
-                    // remove original data
-                    if (uncompressed && autoremove) {
-                        data_file.delete();
+
+                        // remove original data
+                        if (uncompressed && autoremove) {
+                            data_file.delete();
+                        }
                     }
                 }
             }
+            else
+                System.err.println("Skipping " + fileName + " [already exists]");
         }
     }
